requirements.txt
httpx~=0.27.0
nonebot~=1.9.1
aiocqhttp~=1.4.4
pyyaml~=6.0.1
nonebot[scheduler]
ollama~=0.2.1